+++ android/src/main/kotlin/com/ziggeo/flutterplugin/ZiggeoPlugin.kt
@@ -3,10 +3,16 @@ package com.ziggeo.flutterplugin
 import android.os.Handler
 import android.os.Looper
 import com.ziggeo.androidsdk.Ziggeo
+import com.ziggeo.androidsdk.widgets.cameraview.CameraView
+import com.ziggeo.androidsdk.widgets.videoview.ZVideoView
 import com.ziggeo.flutterplugin.api.StreamsMethodChannel
 import com.ziggeo.flutterplugin.api.AudiosMethodChannel
 import com.ziggeo.flutterplugin.api.ImagesMethodChannel
 import com.ziggeo.flutterplugin.api.VideosMethodChannel
+import com.ziggeo.flutterplugin.zrecorder.ZCameraMethodChannel
+import com.ziggeo.flutterplugin.zrecorder.ZCameraRecorderFactory
+import com.ziggeo.flutterplugin.zvideo.ZVideoMethodChannel
+import com.ziggeo.flutterplugin.zvideo.ZVideoViewFactory
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodChannel
 import timber.log.Timber
@@ -22,6 +28,8 @@ class ZiggeoPlugin : FlutterPlugin {
     private lateinit var uplCallbackChannel: MethodChannel
     private lateinit var plCallbackChannel: MethodChannel
     private lateinit var qrCallbackChannel: MethodChannel
+    private lateinit var zvCallbackChannel: MethodChannel
+    private lateinit var zcrCallbackChannel: MethodChannel
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
         ziggeoMainChannel = MethodChannel(binding.binaryMessenger, "ziggeo")
@@ -34,9 +42,22 @@ class ZiggeoPlugin : FlutterPlugin {
         uplCallbackChannel = MethodChannel(binding.binaryMessenger, "ziggeo_upl_callback")
         plCallbackChannel = MethodChannel(binding.binaryMessenger, "ziggeo_pl_callback")
         qrCallbackChannel = MethodChannel(binding.binaryMessenger, "ziggeo_qr_callback")
+        zvCallbackChannel = MethodChannel(binding.binaryMessenger, "z_video_view")
+        zcrCallbackChannel = MethodChannel(binding.binaryMessenger, "z_camera_recorder")
 
         val context = binding.applicationContext
-        val ziggeo = Ziggeo(context)
+        val ziggeo = Ziggeo.getInstance(context)
+        context.setTheme(R.style.ZiggeoTheme)
+        val zVideoView = ZVideoView(context)
+        val zCameraView = CameraView(context)
+
+        binding
+                .platformViewRegistry
+                .registerViewFactory("z_camera_view", ZCameraRecorderFactory(zCameraView))
+        binding
+                .platformViewRegistry
+                .registerViewFactory("z_video_player", ZVideoViewFactory(zVideoView))
+
         ziggeoMainChannel.setMethodCallHandler(
                 ZiggeoMainMethodChannel(ziggeo,
                         recCallbackChannel,
@@ -44,6 +65,7 @@ class ZiggeoPlugin : FlutterPlugin {
                         uplCallbackChannel,
                         plCallbackChannel,
                         qrCallbackChannel,
+                        zvCallbackChannel,
                         Handler(Looper.getMainLooper())
                 )
         )
@@ -51,6 +73,9 @@ class ZiggeoPlugin : FlutterPlugin {
         streamsApiChannel.setMethodCallHandler(StreamsMethodChannel(ziggeo))
         imagesApiChannel.setMethodCallHandler(ImagesMethodChannel(ziggeo))
         audiosApiChannel.setMethodCallHandler(AudiosMethodChannel(ziggeo))
+        zvCallbackChannel.setMethodCallHandler(ZVideoMethodChannel(zVideoView, ziggeo))
+        zcrCallbackChannel.setMethodCallHandler(ZCameraMethodChannel(zCameraView, ziggeo))
+
         Timber.plant(Timber.DebugTree())
     }
 
@@ -65,5 +90,6 @@ class ZiggeoPlugin : FlutterPlugin {
         uplCallbackChannel.setMethodCallHandler(null)
         plCallbackChannel.setMethodCallHandler(null)
         qrCallbackChannel.setMethodCallHandler(null)
+        zvCallbackChannel.setMethodCallHandler(null)
     }
 }