+++ example/lib/screens/auth.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:ziggeo/qr/qr_scanner_config.dart';
 import 'package:ziggeo/qr/qr_scanner_listener.dart';
 import 'package:ziggeo/ziggeo.dart';
 import 'package:ziggeo_example/res/dimens.dart';
-import 'package:ziggeo_example/screens/drawer.dart';
 import 'package:ziggeo_example/screens/main_flow_container.dart';
 import 'package:ziggeo_example/utils/utils.dart';
 import 'package:ziggeo_example/widgets/TextLocalized.dart';
@@ -52,7 +50,8 @@ class _AuthScreenState extends State<AuthScreen> {
     SharedPreferences.getInstance().then((prefs) => prefs
         .setString(Utils.keyAppToken, token)
         .then((success) => Navigator.of(context).pushReplacement(
-            MaterialPageRoute(builder: (context) => MainScreen(token, Ziggeo(token))))));
+            MaterialPageRoute(
+                builder: (context) => MainScreen(token, Ziggeo(token))))));
   }
 
   @override