+++ example/lib/screens/splash.dart
@@ -17,8 +17,8 @@ class _SplashScreenState extends State<SplashScreen> {
     var ziggeo = Ziggeo(appToken);
     if (appToken?.isNotEmpty ?? false) {
       Navigator.of(context).pushReplacement(
-          MaterialPageRoute(builder: (context) => MainScreen(appToken, ziggeo))
-      );
+
+          MaterialPageRoute(builder: (context) => MainScreen(appToken, ziggeo)));
     } else {
       Navigator.of(context).pushReplacement(
           MaterialPageRoute(builder: (context) => AuthScreen()));