+++ lib/recorder/streaming_listener.dart
@@ -4,6 +4,6 @@ typedef OnStreamingStarted = void Function();
 typedef OnStreamingStopped = void Function();
 
 abstract class StreamingEventsListener implements ErrorEventsListener {
-  OnStreamingStarted onStreamingStarted;
-  OnStreamingStopped onStreamingStopped;
+  OnStreamingStarted? onStreamingStarted;
+  OnStreamingStopped? onStreamingStopped;
 }